requirements.txt
pyTelegramBotAPI
requests
python-dotenv
audioread
numpy
scipy
pydub